[NFC] Create `include/mlir-tcp` and move headers over (#12)
This makes it idiomatic to upstream mlir/torch-mlir structure, and makes
downstream integrations less ambiguous in cases like these:
```
- #include "Conversion/Passes.h"
+ #include "mlir-tcp/Conversion/Passes.h"
```
