+++ include/mlir-tcp/Conversion/StablehloToTcp/StablehloToTcp.h
@@ -9,13 +9,16 @@
 
 #pragma once
 
-#include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
+
+#define GEN_PASS_DECL_CONVERTSTABLEHLOTOTCP
+#include "mlir-tcp/Conversion/Passes.h.inc"
+
 namespace tcp {
 
-std::unique_ptr<OperationPass<func::FuncOp>> createConvertTorchToTcpPass();
+std::unique_ptr<Pass> createConvertStablehloToTcpPass();
 
 } // namespace tcp
 } // namespace mlir
