+++ include/mlir-tcp/Conversion/TorchToTcp/TorchToTcp.h
@@ -9,16 +9,17 @@
 
 #pragma once
 
+#include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
 
-#define GEN_PASS_DECL_CONVERTTCPTOARITH
-#include "Conversion/Passes.h.inc"
+#define GEN_PASS_DECL_CONVERTTORCHTOTCP
+#include "mlir-tcp/Conversion/Passes.h.inc"
 
 namespace tcp {
 
-std::unique_ptr<Pass> createConvertTcpToArithPass();
+std::unique_ptr<OperationPass<func::FuncOp>> createConvertTorchToTcpPass();
 
 } // namespace tcp
 } // namespace mlir
