+++ include/mlir-tcp/Dialect/IR/TcpDialect.h
@@ -9,18 +9,7 @@
 
 #pragma once
 
-#include "mlir/Dialect/Func/IR/FuncOps.h"
-#include "mlir/Dialect/Quant/QuantTypes.h"
-#include "mlir/IR/Attributes.h"
-#include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Dialect.h"
 #include "mlir/IR/OpDefinition.h"
-#include "mlir/Interfaces/SideEffectInterfaces.h"
-#include "llvm/ADT/DenseMap.h"
-#include "llvm/ADT/TypeSwitch.h"
 
-#include "Dialect/IR/TcpEnums.h.inc"
-#define GET_ATTRDEF_CLASSES
-#include "Dialect/IR/TcpAttrs.h.inc"
-#define GET_OP_CLASSES
-#include "Dialect/IR/TcpOps.h.inc"
+#include "mlir-tcp/Dialect/IR/TcpDialect.h.inc"
