+++ lib/Conversion/PassDetail.h
@@ -9,7 +9,8 @@
 
 #pragma once
 
-#include "Dialect/IR/TcpDialect.h"
+#include "mlir-tcp/Dialect/IR/TcpDialect.h"
+
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
@@ -18,6 +19,6 @@
 namespace mlir {
 
 #define GEN_PASS_CLASSES
-#include "Conversion/Passes.h.inc"
+#include "mlir-tcp/Conversion/Passes.h.inc"
 
 } // end namespace mlir
