+++ lib/Conversion/TcpToLinalg/Elementwise.cpp
@@ -7,11 +7,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "Conversion/TcpToLinalg/TcpToLinalg.h"
+#include "mlir-tcp/Conversion/TcpToLinalg/TcpToLinalg.h"
+
+#include "mlir-tcp/Dialect/IR/TcpDialect.h"
+#include "mlir-tcp/Dialect/IR/TcpOps.h"
 
 #include "../PassDetail.h"
-#include "Dialect/IR/TcpDialect.h"
-#include "Dialect/IR/TcpOps.h"
 #include "PopulatePatterns.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
