+++ lib/Conversion/TcpToLinalg/TcpToLinalg.cpp
@@ -7,10 +7,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "Conversion/TcpToLinalg/TcpToLinalg.h"
+#include "mlir-tcp/Conversion/TcpToLinalg/TcpToLinalg.h"
 
-#include "Dialect/IR/TcpDialect.h"
-#include "Dialect/IR/TcpOps.h"
+#include "mlir-tcp/Dialect/IR/TcpDialect.h"
+#include "mlir-tcp/Dialect/IR/TcpOps.h"
 
 #include "../PassDetail.h"
 #include "PopulatePatterns.h"
@@ -27,7 +27,7 @@
 namespace mlir {
 
 #define GEN_PASS_DEF_CONVERTTCPTOLINALG
-#include "Conversion/Passes.h.inc"
+#include "mlir-tcp/Conversion/Passes.h.inc"
 
 namespace tcp {
